Replace Nunjucks with LiquidJS, add avatar cache
