+++ .eleventyignore
@@ -1,0 +1,3 @@
+README.md
+CONTRIBUTING-POSTS.md
+_drafts/
