+++ .gitignore
@@ -11,3 +11,4 @@
 _site
 node_modules
 tmp
+.cache
