+++ _includes/filters/regex-replace.js
@@ -8,6 +8,8 @@
  *
  */
 module.exports = (content, rePattern, replacement) => {
+  if (!(replacement && rePattern)) return content;
+
   const re = new RegExp(rePattern, 'g');
 
   return content.replace(re, replacement);
