+++ _includes/nav-posts.liquid
@@ -1,0 +1,13 @@
+<div class="columns is-variable is-8 has-offset-top">
+  <div class="column">
+    {%- if previousPost -%}
+    <a href="{{ previousPost.url | url }}">← {{ previousPost.data.title }}</a>
+    {%- endif -%}
+  </div>
+
+  <div class="column has-text-right">
+    {%- if nextPost -%}
+    <a href="{{ nextPost.url | url }}">{{ nextPost.data.title }} →</a>
+    {%- endif -%}
+  </div>
+</div>
