+++ _includes/shortcodes/meetup-details.js
@@ -1,6 +1,22 @@
 const organizations = require('../../_data/organizations.json');
 const siteData = require('../../_data/site.json');
-const filterFullDate = require('../filters/full-date');
+
+/**
+ * Filter date output to follow pattern MMMM D, YYYY
+ *
+ * @param  {object}   dateValue  Date object
+ *
+ * @return {string}   Formatted date string.
+ */
+const fullDate = (dateValue) => {
+  const monthNames = [
+    'January', 'February', 'March', 'April', 'May', 'June',
+    'July', 'August', 'September', 'October', 'November', 'December',
+  ];
+  const newDate = new Date(dateValue);
+
+  return `${monthNames[newDate.getMonth()]} ${newDate.getDate()}, ${newDate.getFullYear()}`;
+};
 
 /**
  * Provide template for meeting details.
@@ -58,7 +74,7 @@ module.exports = function meetupDetails(meetDate, venue, after, msgHeader, meetT
       <span class="icon has-text-grey-dark">
         <ion-icon src="/assets/ionicon/calendar.svg"></ion-icon>
       </span>
-      <time class="has-text-primary has-text-weight-bold" dateTime="${meetDate.toISOString().substring(0, 10)}">${filterFullDate(meetDate)}</time>
+      <time class="has-text-primary has-text-weight-bold" dateTime="${meetDate.toISOString().substring(0, 10)}">${fullDate(meetDate)}</time>
     </div>
 
     <div>
