+++ _includes/shortcodes/video-player.js
@@ -16,6 +16,11 @@ module.exports = (video, title) => {
       } else if (videoMatch[1] === 'youtu.be') {
         videoPlayer = `<iframe class="has-ratio" width="640" height="380" src="https://www.youtube.com/embed/${videoId}" frameborder="0" title="{{ title }}" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>`;
         videoSrc = ' on YouTube';
+      } else if (videoMatch[1] === 'www.youtube.com') {
+        const videoVId = videoId.replace(/.*v=([a-z0-9]+).*/i, '$1');
+
+        videoPlayer = `<iframe class="has-ratio" width="640" height="380" src="https://www.youtube.com/embed/${videoVId}" frameborder="0" title="{{ title }}" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>`;
+        videoSrc = ' on YouTube';
       } else {
         return `<div class="message flex-item is-info">
           <h3 class="message-header">Meetup video</h3>
