+++ _layouts/meetup.liquid
@@ -8,84 +8,72 @@ layout: base
       {{ content | safe }}
     </div>
 
-    {% meetupDetails page.date, venue, after %}
+    {%- meetupDetails page.date, venue, after -%}
 
-    {% videoPlayer video, title %}
+    {%- videoPlayer video, title -%}
   </div>
 
   <div class="column has-offset-top">
-    {% if slides %}
+    {%- if slides -%}
     <div class="message flex-item is-info">
       <h3 class="message-header">Meetup slides</h3>
       <ul class="message-body">
-        {% if slides %}
+        {%- if slides -%}
         <li><a href="{{ slides }}">Get the slide deck!</a></li>
-        {% endif %}
+        {%- endif -%}
       </ul>
     </div>
 
-    {% endif %}
+    {%- endif -%}
     <h2 class="title is-sr-only">Speaker</h2>
 
-    {% for person in speakers %}
+    {%- for person in speakers -%}
     <div class="has-offset-top">
-      {% include 'business-card.njk' %}
+      {%- include business-card.liquid -%}
     </div>
-    {% endfor%}
+    {%- endfor -%}
   </div>
 </div>
 
-{% if sponsor.name or sponsor %}
-{% if sponsor.name and sponsor.url %}
-{% set sponsorOrg = sponsor %}
-{% elif sponsor %}
-{% set sponsorOrg = organizations[sponsor] %}
-{% endif %}
+{%- if sponsor.name or sponsor -%}
+{%- if sponsor.name and sponsor.url -%}
+{%- assign sponsorOrg = sponsor -%}
+{%- elsif sponsor -%}
+{%- assign sponsorOrg = organizations[sponsor] -%}
+{%- endif -%}
 <div class="columns is-variable is-8-desktop">
   <div class="column is-two-thirds">
     <div class="box is-8 content">
       <h3>Sponsor</h3>
 
-      {% if sponsorOrg.logo %}
+      {%- if sponsorOrg.logo -%}
       <figure class="image sponsor-logo">
         <img src="{{ sponsorOrg.logo }}" alt="Sponsored by: {{ sponsorOrg.name }}">
       </figure>
-      {% endif %}
+      {%- endif -%}
 
-      {% if sponsorOrg.message %}
+      {%- if sponsorOrg.message -%}
       <div>
         {{ sponsorOrg.message | markdown | safe }}
       </div>
-      {% elif sponsorOrg.name %}
+      {%- elsif sponsorOrg.name -%}
       <p>Austin JavaScript is <strong>sponsored this month by <a href="{{ sponsorOrg.url }}">{{ sponsorOrg.name }}</a></strong>. Please thank them for their gracious gifts of pizza and drinks and for their support of the local JavaScript community.
-        {% if sponsorOrg.careerUrl %}
+        {%- if sponsorOrg.careerUrl -%}
         Ask them about their <a href="{{ sponsorOrg.careerUrl }}">job opportunities</a>!
-        {% endif %}
+        {%- endif -%}
       </p>
-      {% endif %}
+      {%- endif -%}
 
-      {% if sponsorOrg.note %}
+      {%- if sponsorOrg.note -%}
       <p>
         {{ sponsorOrg.note | markdown | safe }}
       </p>
-      {% endif %}
+      {%- endif -%}
     </div>
   </div>
 </div>
-{% endif %}
+{%- endif -%}
 
-<div class="columns is-variable is-8 has-offset-top">
-  {% set nextPost = collections.meetups | getNextCollectionItem(page) %}
-  {% set previousPost = collections.meetups | getPreviousCollectionItem(page) %}
-  <div class="column">
-    {% if previousPost %}
-    <a href="{{ previousPost.url | url }}">← {{ previousPost.data.title }}</a>
-    {% endif %}
-  </div>
-
-  <div class="column has-text-right">
-    {% if nextPost %}
-    <a href="{{ nextPost.url | url }}">{{ nextPost.data.title }} →</a>
-    {% endif %}
-  </div>
-</div>
+{%- assign nextPost = collections.meetups | getNextCollectionItem: page -%}
+{%- assign previousPost = collections.meetups | getPreviousCollectionItem: page -%}
+{%- include nav-posts.liquid -%}
