+++ _meetups/2015-06-16-meetup.md
@@ -25,7 +25,7 @@ organizers:
   - astacy
 ---
 
-Howdy, Texas. We're looking forward to seeing all of you fine vaqueras and vaqueros this Tuesday {{ when | fullDate }} as Kyle Simpson ([@getify][]) rustles up some knowledge on transpiling JavaScript.
+Howdy, Texas. We're looking forward to seeing all of you fine vaqueras and vaqueros this Tuesday {{ when | date: '%B %e, %Y' }} as Kyle Simpson ([@getify][]) rustles up some knowledge on transpiling JavaScript.
 
 The new reality of JavaScript is that the features will evolve even quicker than the specification, and much quicker than your "supported" browsers. So we're going to have to come to grips with transpilers being a standard part of our build processes. But what can we do to wrangle at least two versions of every file? What does that mean for server-side coding (node/iojs) and what does it mean for browser-delivered files? Should we use the transpiled code everywhere, or should we have split delivery?
