+++ _meetups/2015-07-21-meetup.md
@@ -36,7 +36,7 @@ organizers:
 
 Sup, y'all. We've got something special planned for our July meetup and we are _shore_ you're gonna think it's fresh to death.
 
-Here's [the Situation][the-situation]: AustinJS is going down the Shore! Jenn Schiffer is staging a blowout [JerseyScript][jerseyscript] bo-coup on {{ when | fullDate }}. Jenn has convinced New York's own Adam J. Sontag to come rap with us about a speech recognition game he is working on.
+Here's [the Situation][the-situation]: AustinJS is going down the Shore! Jenn Schiffer is staging a blowout [JerseyScript][jerseyscript] bo-coup on {{ when | date: '%B %e, %Y' }}. Jenn has convinced New York's own Adam J. Sontag to come rap with us about a speech recognition game he is working on.
 
 [the-situation]: https://en.wikipedia.org/wiki/Michael_Sorrentino
 [jerseyscript]: http://jerseyscript.github.io/
