+++ _meetups/2015-08-18-meetup.md
@@ -25,7 +25,7 @@ organizers:
 We're not in our usual location! We'll be meeting at <strong>Spredfast</strong> this month in the Silicon Labs building, so please plan accordingly.
 </div>
 
-Hey Austin JavaScripters, summer is in full swing and `{% raw %}{{ joke_about_how_ember_is_hot }}{% endraw %}`. We're excited to hear from [Iheanyi Ekechukwu][] as he tells us about "Ember.js, Ember-CLI, and how it all relates to that 'shut up and dance with me' song," (which we were actually [hoping to hear about][] last month). Hope to see you there this Tuesday, {{ when | fullDate }}.
+Hey Austin JavaScripters, summer is in full swing and `{% raw %}{{ joke_about_how_ember_is_hot }}{% endraw %}`. We're excited to hear from [Iheanyi Ekechukwu][] as he tells us about "Ember.js, Ember-CLI, and how it all relates to that 'shut up and dance with me' song," (which we were actually [hoping to hear about][] last month). Hope to see you there this Tuesday, {{ when | date: '%B %e, %Y' }}.
 
 [iheanyi ekechukwu]: http://www.iheanyi.com/
 [hoping to hear about]: https://twitter.com/fivetanley/status/623620046338666497
