+++ _meetups/2016-02-16-meetup.md
@@ -26,6 +26,6 @@ organizers:
 
 > **We're not in our usual location!**: We'll be meeting at Spredfast this month in the Silicon Labs Building on Colorado/Cesar Chavez ([map](https://www.google.com/maps/place/200+W+Cesar+Chavez+St,+Austin,+TX+78701/@30.2642656,-97.7470567,18z/data=!3m1!4b1!4m2!3m1!1s0x8644b50602c5b57d:0x4c4d44de892b1d04)).
 
-Hey everyone, we're excited about {{ when | fullDate }} because Benjamin Coe is going to come talk about _{{ title }}_.
+Hey everyone, we're excited about {{ when | date: '%B %e, %Y' }} because Benjamin Coe is going to come talk about _{{ title }}_.
 
 We've got food and beverages this month thanks to **Front End Development at IBM**. If you'd like to hear more about what they're working on, feel free to reach out to frontend@us.ibm.com or check out the [Feducation event they're doing with General Assembly](https://generalassemb.ly/education/ga-fedibm-design-present-the-importance-of-a-pattern-library/austin/21730).
