+++ _meetups/2016-04-19-meetup.md
@@ -24,6 +24,6 @@ organizers:
   - astacy
 ---
 
-It's April, so let's shake off our collective SXSW hangovers 🍻 (or maybe emerge from our SXSW bunkers ☢️) and get together on {{ when | fullDate }} for another great AustinJS!
+It's April, so let's shake off our collective SXSW hangovers 🍻 (or maybe emerge from our SXSW bunkers ☢️) and get together on {{ when | date: '%B %e, %Y' }} for another great AustinJS!
 
 We'll be hearing from Anton Astashov about [Elm](http://elm-lang.org/), a transpiled-to-JavaScript language which makes sure you’ll never have a runtime error in your app. Almost. He'll cover why it’s cool, what it looks like, it’s pros and (most importantly) cons!
