+++ _meetups/2016-05-17-meetup.md
@@ -25,4 +25,4 @@ organizers:
   - astacy
 ---
 
-This month we've got [dark web expert](https://miketaylr.com/posts/2016/04/string-prototype-contains-use-your-own-judgement.html) and [Gotye fan](https://miketaylr.com/posts/2016/02/dispatching-magical-webkit-prefixed-events.html) [<strike>Miike</strike> Mike <strike>Snow</strike> Taylor]({{ twiturl }}) in to discuss the nuances of writing JavaScript that actually works in world-wide web browsing machines. Come out and join us next Tuesday {{ when | fullDate }}!
+This month we've got [dark web expert](https://miketaylr.com/posts/2016/04/string-prototype-contains-use-your-own-judgement.html) and [Gotye fan](https://miketaylr.com/posts/2016/02/dispatching-magical-webkit-prefixed-events.html) [<strike>Miike</strike> Mike <strike>Snow</strike> Taylor]({{ twiturl }}) in to discuss the nuances of writing JavaScript that actually works in world-wide web browsing machines. Come out and join us next Tuesday {{ when | date: '%B %e, %Y' }}!
