+++ _meetups/2016-06-21-meetup.md
@@ -21,4 +21,4 @@ organizers:
   - astacy
 ---
 
-This month we've got veteran frontend engineer Rebecca Murphey, who will share the story of how she started a new job this year by paying a visit to JavaScript circa 2009, back when Ryan Dahl was getting ready to announce Node and Facebook was still four years away from being mocked for the apparent heresy of JSX. She'll explain how she modernized and best-practice-ified a project that didn't even have a package.json, smoothing the development process, eliminating common sources of bugs, paving the way for bigger improvements, and never once uttering the words "we oughta just start from scratch." Come out and join us next Tuesday {{ when | fullDate }}!
+This month we've got veteran frontend engineer Rebecca Murphey, who will share the story of how she started a new job this year by paying a visit to JavaScript circa 2009, back when Ryan Dahl was getting ready to announce Node and Facebook was still four years away from being mocked for the apparent heresy of JSX. She'll explain how she modernized and best-practice-ified a project that didn't even have a package.json, smoothing the development process, eliminating common sources of bugs, paving the way for bigger improvements, and never once uttering the words "we oughta just start from scratch." Come out and join us next Tuesday {{ when | date: '%B %e, %Y' }}!
