+++ _meetups/2016-07-19-meetup.md
@@ -21,4 +21,4 @@ organizers:
   - astacy
 ---
 
-This Tuesday, {{ when | fullDate }}, discover how this one weird extension **_changed the web forever_**. Andrew Dupont will give us a 12 year perspective (most of a century in internet years) on the ways GreaseMonkey shaped today's browsers and how user scripts give us the opportunity of a better web.
+This Tuesday, {{ when | date: '%B %e, %Y' }}, discover how this one weird extension **_changed the web forever_**. Andrew Dupont will give us a 12 year perspective (most of a century in internet years) on the ways GreaseMonkey shaped today's browsers and how user scripts give us the opportunity of a better web.
