+++ _meetups/2016-08-16-meetup.md
@@ -59,4 +59,4 @@ organizers:
   100% {-webkit-transform: scale(0.9,1.0);transform: scale(0.9,1.0);}
 }</style>
 
-Come out next Tuesday, {{ when | fullDate }}, as Dave Rupert tells us how to <strike>make the mobile web great again</strike> make <span class="wowowow">Progressive Web Apps!</span>
+Come out next Tuesday, {{ when | date: '%B %e, %Y' }}, as Dave Rupert tells us how to <strike>make the mobile web great again</strike> make <span class="wowowow">Progressive Web Apps!</span>
