+++ _meetups/2016-11-15-meetup.md
@@ -21,7 +21,7 @@ organizers:
   - astacy
 ---
 
-Hey happy Veteran's Day and Marine Corps Birthday! AustinJS is going to get together Tuesday {{ when | fullDate }} to hear from Jon Loyens about data-driven front end development `~⭐️for the troops⭐️~`
+Hey happy Veteran's Day and Marine Corps Birthday! AustinJS is going to get together Tuesday {{ when | date: '%B %e, %Y' }} to hear from Jon Loyens about data-driven front end development `~⭐️for the troops⭐️~`
 
 In his own words:
