+++ _meetups/2017-07-18-meetup.md
@@ -22,7 +22,7 @@ organizers:
   - astacy
 ---
 
-It's too hot for hair out there, so we've asked Lon Ingram to come talk to us about making websites fast as &#9889; (kung fu fighting, etc). Join us Tuesday {{ when | fullDate }} and learn what it takes to appease our search engine overlords!
+It's too hot for hair out there, so we've asked Lon Ingram to come talk to us about making websites fast as &#9889; (kung fu fighting, etc). Join us Tuesday {{ when | date: '%B %e, %Y' }} and learn what it takes to appease our search engine overlords!
 
 ## Talk summary
