+++ _meetups/2017-08-15-meetup.md
@@ -22,7 +22,7 @@ organizers:
   - astacy
 ---
 
-Ever wonder how to pronounce Babel? Join us Tuesday {{ when | fullDate }} to find out! Austin's own Andrew Levine will teach us how to write Babel plugins.
+Ever wonder how to pronounce Babel? Join us Tuesday {{ when | date: '%B %e, %Y' }} to find out! Austin's own Andrew Levine will teach us how to write Babel plugins.
 
 ## Talk summary
