+++ _meetups/2017-09-19-meetup.md
@@ -25,7 +25,7 @@ organizers:
 
 Source maps, huh, yeah. What are they good for? Absolutely... some things.
 
-We have an extra special treat this month thanks to our gracious hosts and good friends at Spredfast: Ben Vinegar is coming all the way from "San Francisco" to tell us everything about how source maps work. Join us Tuesday {{ when | fullDate }} for a special Bae Area AustinJS!
+We have an extra special treat this month thanks to our gracious hosts and good friends at Spredfast: Ben Vinegar is coming all the way from "San Francisco" to tell us everything about how source maps work. Join us Tuesday {{ when | date: '%B %e, %Y' }} for a special Bae Area AustinJS!
 
 ## Talk summary
