+++ _meetups/2018-05-15-meetup.md
@@ -12,7 +12,7 @@ organizers:
   - astacy
 ---
 
-Come out tomorrow {{ when | fullDate }}, for some `bevera.js` at the Gingerman! Topics of discussion may include:
+Come out tomorrow {{ when | date: '%B %e, %Y' }}, for some `bevera.js` at the Gingerman! Topics of discussion may include:
 
 * The open web
 * [JorbaScrump](https://twitter.com/davatron5000/status/369187413291065344)
