+++ posts/meetups.liquid
@@ -1,32 +1,32 @@
 ---
 layout: base
-title: Posts archive
+title: Meetup archive
 meta:
-  description: An archive of past Austin JavaScript posts and meetups.
+  description: An archive of past Austin JavaScript meetups.
 ---
 <div class="content">
   <h1>{{ title }}</h1>
   <p>{{ meta.description }}</p>
 
   <ul>
-    {% for post in collections.posts | reverse %}
+    {%- assign meetups = collections.meetups -%}
+    {%- for post in meetups -%}
     <li>
       <a href="{{ post.url | url }}">
-        {{ post.inputPath.replace('./_meetups/', '').replace('./_posts/', '').substring(0, 10) }}:
+        {{ post.inputPath | slice: 11, 10 }}:
         <span class="has-text-weight-semibold">{{ post.data.title }}</span>
-        {{ '[post]' if post.fileSlug != 'meetup' }}
-        {% if post.data.video %}
+        {%- if post.data.video -%}
         <span class="icon tooltip" title="video available">
           <ion-icon src="{{ '/assets/ionicon/videocam.svg' | url }}"></ion-icon>
         </span>
-        {% endif %}
-        {% if post.data.slides %}
+        {%- endif -%}
+        {%- if post.data.slides -%}
         <span class="icon tooltip" title="slides available">
           <ion-icon src="{{ '/assets/ionicon/albums.svg' | url }}"></ion-icon>
         </span>
-        {% endif %}
+        {%- endif -%}
       </a>
     </li>
-    {% endfor %}
+    {%- endfor -%}
   </ul>
 </div>
